Adding in clojure helper and fillout out interfaces.

+++ java/tech/v3/datatype/LongReader.java
@@ -3,6 +3,6 @@ package tech.v3.datatype;
 
 public interface LongReader extends LongBuffer
 {
-  default void writeLong(long idx, byte value) { throw new UnsupportedOperationException(); }
+  default void writeLong(long idx, long value) { throw new UnsupportedOperationException(); }
   default boolean supportsWrite() { return false; }
 }
